// src/modules/link/interfaces/link.interface.ts
export enum LinkStatus {
  PENDENTE = 'PENDENTE',
  RESPONDIDO = 'RESPONDIDO',
  EXPIRADO = 'EXPIRADO',
}

export enum CanalEnvio {
  WHATSAPP = 'WHATSAPP',
  EMAIL = 'EMAIL',
  MANUAL = 'MANUAL',
}

export enum MotivoInvalido {
  EXPIRADO = 'EXPIRADO',
  JA_RESPONDIDO = 'JA_RESPONDIDO',
  NAO_ENCONTRADO = 'NAO_ENCONTRADO',
}